Move Aircraft class effects to Aircraft_base.py
